+++ src/etables.c
@@ -64,6 +64,7 @@ static bool route = false;
 static bool passthru = false;
 static bool intercept = false;
 static bool echo = false;
+static bool verbose = false;
 static bool disable = false;
 static struct in_addr dcidr;
 static struct in_addr scidr;
@@ -78,7 +79,8 @@ static unsigned short protocol;
 static const char *path = "/sys/fs/bpf/tc/globals/zt_tproxy_map";
 static char doc[] = "etables -- ebpf mapping tool";
 static char *echo_interface;
-const char *argp_program_version = "0.2.7";
+static char *verbose_interface;
+const char *argp_program_version = "0.2.8";
 int get_key_count();
 
 struct ifindex_ip4
@@ -87,8 +89,9 @@ struct ifindex_ip4
     char ifname[IF_NAMESIZE];
 };
 
-struct icmp_ip4 {
+struct diag_ip4 {
     bool echo;
+    bool verbose;
 };
 
 struct tproxy_port_mapping
@@ -374,46 +377,70 @@ void usage(char *message)
     fprintf(stderr, "       etables -L -i\n");
     fprintf(stderr, "       etables -L -f\n");
     fprintf(stderr, "       etables -V\n");
+    fprintf(stderr, "       etables -e <ifname>\n");
+    fprintf(stderr, "       etables -e <ifname> -d\n");
+    fprintf(stderr, "       etables -v <ifname>\n");
+    fprintf(stderr, "       etables -v <ifname> -d\n");
     fprintf(stderr, "       etables --help\n");
     exit(1);
 }
 
-bool set_echo(int *idx){
+bool set_diag(int *idx){
     /* create bpf_attr to store ifindex_ip_map */
-    union bpf_attr echo_map;
+    union bpf_attr diag_map;
     /*path to pinned ifindex_ip_map*/
-    const char *icmp_map_path = "/sys/fs/bpf/tc/globals/icmp_map";
+    const char *diag_map_path = "/sys/fs/bpf/tc/globals/diag_map";
     /* open BPF ifindex_ip_map */
-    memset(&echo_map, 0, sizeof(echo_map));
+    memset(&diag_map, 0, sizeof(diag_map));
     /* set path name with location of map in filesystem */
-    echo_map.pathname = (uint64_t)icmp_map_path;
-    echo_map.bpf_fd = 0;
-    echo_map.file_flags = 0;
+    diag_map.pathname = (uint64_t)diag_map_path;
+    diag_map.bpf_fd = 0;
+    diag_map.file_flags = 0;
     /* make system call to get fd for map */
-    int icmp_fd = syscall(__NR_bpf, BPF_OBJ_GET, &echo_map, sizeof(echo_map));
+    int icmp_fd = syscall(__NR_bpf, BPF_OBJ_GET, &diag_map, sizeof(diag_map));
     if (icmp_fd == -1)
     {
         printf("BPF_OBJ_GET: %s \n", strerror(errno));
         exit(1);
     }
-    echo_map.map_fd = icmp_fd;
-    struct icmp_ip4 if_icmp = {0};
-    if(!disable){
-        if_icmp.echo = true;
-    }
-    else{
-        if_icmp.echo = false;
-    }     
-    echo_map.key = (uint64_t)idx;
-    echo_map.flags = BPF_ANY;
-    echo_map.value = (uint64_t)&if_icmp;
-    int ret = syscall(__NR_bpf, BPF_MAP_UPDATE_ELEM, &echo_map, sizeof(echo_map));
+    diag_map.map_fd = icmp_fd;
+    struct diag_ip4 o_diag;
+    diag_map.key = (uint64_t)idx;
+    diag_map.flags = BPF_ANY;
+    diag_map.value = (uint64_t)&o_diag;
+    int lookup = syscall(__NR_bpf, BPF_MAP_LOOKUP_ELEM, &diag_map, sizeof(diag_map));
+    if(lookup){
+        printf("Invalid Index\n");
+        close(icmp_fd);
+        return false;
+    }else{
+        if(echo){
+            if(!disable){
+                o_diag.echo = true;
+            }
+            else{
+                o_diag.echo = false;
+            }
+             printf("Set icmp-echo to %d for %s\n", !disable, echo_interface);
+        }
+        if(verbose){
+            if(!disable){
+                o_diag.verbose = true;
+            }
+            else{
+                o_diag.verbose = false;
+            }
+            printf("Set verbose to %d for %s\n", !disable, verbose_interface);
+        }
+    }  
+    int ret = syscall(__NR_bpf, BPF_MAP_UPDATE_ELEM, &diag_map, sizeof(diag_map));
     if (ret)
     {
         printf("MAP_UPDATE_ELEM: %s \n", strerror(errno));
         close(icmp_fd);
         exit(1);
     }
+    close(icmp_fd);
     return true;
 }
 
@@ -457,6 +484,7 @@ bool interface_map()
     in_addr_t ifip;
     int ipcheck = 0;
     bool create_route = true;
+    int lo_count = 0;
     while (address)
     {
         if (address->ifa_addr && (address->ifa_addr->sa_family == AF_INET))
@@ -477,6 +505,7 @@ bool interface_map()
             else
             {
                ifip = 0x0100007f;
+               lo_count++;
             }
             struct ifindex_ip4 ifip4 = {
                 ifip,
@@ -485,11 +514,16 @@ bool interface_map()
             if_map.key = (uint64_t)&idx;
             if_map.flags = BPF_ANY;
             if_map.value = (uint64_t)&ifip4;
-            if(echo){
+            if(echo && !(idx == 1)){
                 if(!strcmp(echo_interface, address->ifa_name)){
-                    if(set_echo(&idx)){
-                        printf("successfully changed icmp echo for %s to %d\n", address->ifa_name, !disable);
-                    }
+                    set_diag(&idx);
+                }
+            }else if(echo && !strcmp(echo_interface,"lo") && (idx == 1) && lo_count == 1){
+                printf("icmp echo is always set to 1 for lo\n");
+            }
+            if(verbose && !((idx==1) && lo_count > 1)){
+                if(!strcmp(verbose_interface, address->ifa_name)){
+                    set_diag(&idx);
                 }
             }
             int ret = syscall(__NR_bpf, BPF_MAP_UPDATE_ELEM, &if_map, sizeof(if_map));
@@ -923,6 +957,7 @@ static struct argp_option options[] = {
     {"flush", 'F', NULL, 0, "Flush all map rules", 0},
     {"dcidr-block", 'c', "", 0, "Set dest ip prefix i.e. 192.168.1.0 <mandatory for insert/delete/list>", 0},
     {"icmp-echo", 'e', "", 0, "enable inbound icmp echo to interface", 0},
+    {"verbose", 'v', "", 0, "enable inbound icmp echo to interface", 0},
     {"disable", 'd', NULL, 0, "disabble associated icmp echo operation i.e. -e eth0 -d to disable inbound echo on eth0", 0},
     {"ocidr-block", 'o', "", 0, "Set origin ip prefix i.e. 192.168.1.0 <mandatory for insert/delete/list>", 0},
     {"dprefix-len", 'm', "", 0, "Set dest prefix length (1-32) <mandatory for insert/delete/list >", 0},
@@ -1031,6 +1066,16 @@ static error_t parse_opt(int key, char *arg, struct argp_state *state)
         tproxy_port = port2s(arg);
         tpt = true;
         break;
+    case 'v':
+        if (!strlen(arg))
+        {
+            fprintf(stderr, "Interface name required as arg to -v, --verbose: %s\n", arg);
+            fprintf(stderr, "%s --help for more info\n", program_name);
+            exit(1);
+        }
+        verbose = true;
+        verbose_interface = arg;
+        break;
     default:
         return ARGP_ERR_UNKNOWN;
     }
@@ -1043,8 +1088,25 @@ int main(int argc, char **argv)
 {
     argp_parse(&argp, argc, argv, 0, 0, 0);
 
-    if(disable && !echo){
-        usage("Missing argument -e, --icmp-echo");
+    if(echo && verbose){
+        usage("Can not set -e, --icmp-echo and -v, --verbose with single call to etables");
+    }
+
+    if(echo && (add || delete || list || flush)){
+        usage("Can not set -e, --icmp-echo with any other calls to etables besides -d");
+    }    
+
+    if(verbose && (add || delete || list || flush)){
+        usage("Can not set -v, --verbose with any other calls to etables besides -d");
+    }   
+
+    if(disable && (!echo && !verbose)){
+        usage("Missing argument at lease one of -e, --icmp-echo or -v, --verbose");
+    }
+
+    if(verbose){
+        interface_map();
+        exit(0);
     }
 
     if(echo){